+++ src/kernel/config.h
@@ -284,7 +284,7 @@ extern "C" {
     const char *get_param(const struct param *p, const char *key);
     int get_param_int(const struct param *p, const char *key, int def);
     int check_param(const struct param *p, const char *key, const char *searchvalue);
-    float get_param_flt(const struct param *p, const char *key, float def);
+    double get_param_flt(const struct param *p, const char *key, double def);
     void free_params(struct param **pp);
 
     bool ExpensiveMigrants(void);
